clippy fix

+++ examples/split.rs
@@ -153,5 +153,5 @@ fn update_camera_2(
     camera_driver.rotate_yaw_pitch(-1.0, 0.0);
 
     let a = rig.driver_mut::<Arm>();
-    a.offset = Vec3::Z * ((time.delta_seconds() as f32 * 0.2).sin().cos().abs() * 400. - 200.);
+    a.offset = Vec3::Z * ((time.delta_seconds() * 0.2).sin().cos().abs() * 400. - 200.);
 }
